+++ SistemaAcademia/SistemaAcademia/Models/ClienteViewModel.cs
@@ -7,7 +7,7 @@ public class CadastroClienteViewModel
     public DateTime DataNascimento { get; set; }
     public string Email { get; set; } = string.Empty;
     public string Senha { get; set; } = string.Empty;
-    public int IdOrientacaoSexual { get; set; }
+    public long IdOrientacaoSexual { get; set; }
     public string Telefone { get; set; } = string.Empty;
     public string TelefoneOpcional { get; set; } = string.Empty;
     public string CEP { get; set; } = string.Empty;
@@ -19,16 +19,16 @@ public class CadastroClienteViewModel
     public string Estado { get; set; } = string.Empty;
 
     // matrícula opcional — preenchida na mesma tela de cadastro
-    public int?      IdModalidade               { get; set; }
-    public int?      IdPlano                    { get; set; }
+    public long?     IdModalidade               { get; set; }
+    public long?     IdPlano                    { get; set; }
     public DateTime? DataInicioMatricula         { get; set; }
     public bool      PrimeiroPagoNoAto           { get; set; }
-    public int?      IdMetodoPagamentoPrimeiro   { get; set; }
+    public long?     IdMetodoPagamentoPrimeiro   { get; set; }
 }
 
 public class ClienteListaViewModel
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Nome { get; set; } = string.Empty;
     public string CPF { get; set; } = string.Empty;
     public string Email { get; set; } = string.Empty;
@@ -39,12 +39,12 @@ public class ClienteListaViewModel
 
 public class EditarClienteViewModel
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string CPF { get; set; } = string.Empty;
     public string Nome { get; set; } = string.Empty;
     public string Email { get; set; } = string.Empty;
     public DateTime DataNascimento { get; set; }
-    public int IdOrientacaoSexual { get; set; }
+    public long IdOrientacaoSexual { get; set; }
     public string CEP { get; set; } = string.Empty;
     public string Logradouro { get; set; } = string.Empty;
     public string Numero { get; set; } = string.Empty;
