+++ SistemaAcademia/SistemaAcademia/Models/Plano.cs
@@ -1,10 +1,10 @@
 namespace SistemaAcademia.Models;
 
-public class Plano(string nome, string tempoPlano, float percentualDesconto, int idAcademia)
+public class Plano(string nome, string tempoPlano, float percentualDesconto, long idAcademia)
 {
-    public int    Id                 { get; init; }
+    public long   Id                 { get; init; }
     public string Nome               { get; } = nome;
     public string TempoPlano         { get; } = tempoPlano;
     public float  PercentualDesconto { get; } = percentualDesconto;
-    public int    IdAcademia         { get; } = idAcademia;
+    public long   IdAcademia         { get; } = idAcademia;
 }
